+++ cmd/harbor/root/project/cmd.go
@@ -1,7 +1,6 @@
 package project
 
 import (
-	"github.com/goharbor/harbor-cli/cmd/harbor/root/project/metadata"
 	"github.com/spf13/cobra"
 )
 
@@ -18,7 +17,7 @@ func Project() *cobra.Command {
 		ListProjectCommand(),
 		ViewCommand(),
 		LogsProjectCommmand(),
-		metadata.Metadata(),
+		ProjectMetadataCommand(),
 	)
 
 	return cmd
